+++ jklib/dj/models.py
@@ -10,31 +10,23 @@ from django.utils.deconstruct import deconstructible
 
 
 class ImprovedModel(models.Model):
-
     class Meta:
         abstract = True
 
     def save(self, *args: Any, **kwargs: Any) -> None:
-        is_create = self.pk is None
-        self._pre_create() if is_create else self._pre_update()
+        self._pre_save()
         super().save(*args, **kwargs)
-        self._post_create() if is_create else self._post_update()
+        self._post_save()
 
     def delete(self, *args: Any, **kwargs: Any) -> None:
         self._pre_delete()
         super().delete(*args, **kwargs)
         self._post_delete()
 
-    def _pre_create(self) -> None:
+    def _pre_save(self) -> None:
         pass
 
-    def _post_create(self) -> None:
-        pass
-
-    def _pre_update(self) -> None:
-        pass
-
-    def _post_update(self) -> None:
+    def _post_save(self) -> None:
         pass
 
     def _pre_delete(self) -> None: